+++ CHANGELOG.md
@@ -9,6 +9,12 @@
 * `Filesystem::isWritableResursive()` has been removed.
 
 ## 1.6 branch
+### 1.6.5
+* `isReadable()` and `isWritable()` method provided by `Exceptionist` no longer even call the
+  `fileExists()` method;
+* added `\Tools\TestSuite\TestTrait::assertDeprecated()` method;
+* improved `\Tools\TestSuite\TestTrait::assertException()`.
+
 ### 1.6.4
 * `object_map()` global function and `\Tools\Exceptionist::methodExists()` method throw the new 
   `\Tools\Exception\MethodNotExistsException`;